Refactor | ReservationPaymentManager, enums changed

+++ hotel-service/src/main/java/com/ossovita/hotelservice/business/abstracts/RoomService.java
@@ -1,6 +1,5 @@
 package com.ossovita.hotelservice.business.abstracts;
 
-import com.ossovita.commonservice.core.entities.enums.RoomStatus;
 import com.ossovita.hotelservice.core.entities.Room;
 import com.ossovita.hotelservice.core.entities.dto.request.RoomRequest;
 
@@ -14,7 +13,7 @@ public interface RoomService {
 
     boolean isRoomAvailable(long roomPk);
 
-    int getRoomPriceWithRoomFk(long roomFk);
+    double getRoomPriceWithRoomFk(long roomFk);
 
     List<Room> getAvailableRoomsByHotelFk(long hotelFk);
